requirements.txt
Django==2.0.13
django-filter==2.1.0
django-multiselectfield==0.1.8
pkg-resources==0.0.0
pytz==2018.9
Pillow==5.4.1
djangorestframework==3.9.2

